big refactor of resources
no more passing in VARNAME_ID as args, now it's using the
manage_relationships stuff properly

haven't built out the loop stuff yet, so there might be a couple more
required changes there

also refactor the test generators to just generate the resources (not
just changesets which will be piped into generate/1)

the change to the generators stuff is because I'm still not 100% sure if
the Ash.Generators module is working right - the docs seem to be a bit
incomplete in a couple of places? Or perhaps it's that I'm holding it
wrong (entirely possible). Either way, this works for now and we can come
back to it later