refactor variable voting component
this change is to have it toggle between not completed/completed without changing
the URL. It was super tricky to get it working with a stream of :variables (seriously,
don't ask me how long I spent on that or I'll cry). Now it's just using a normal
list of :variables and it's easy.

one day I might return to this and figure out how to get it working with streams