add require_atomic? false to several actions
this is basically to make it work again (Postgres seems stricter
than SQLite in this regard)

there was also a change to the create session/add participant interplay, but
that was fixed by doing an after_transaction instead of after_action.
A bit gross, but not likely to cause problems in practice.

Finally, the "loop sorting" code might be broken again (or maybe it never actually)
worked? Need to investigate more. For now, one test is still broken (but
the rest all pass).